tools/requirements.txt
mysql-connector-python==8.2.0
Faker==20.1.0
bcrypt==4.1.2



